+++ nix/flake-module.nix
@@ -32,7 +32,13 @@ in
                 '';
                 default = false;
               };
-
+              drv = mkOption {
+                type = types.package;
+                readOnly = true;
+                description = ''
+                  The `hlsCheck` derivation generated for this project.
+                '';
+              };
             };
           };
           packageSubmodule = with types; submodule {
@@ -136,13 +142,14 @@ in
                   The development shell derivation generated for this project.
                 '';
               };
-              hlsCheck = mkOption {
-                type = types.package;
+              checks = mkOption {
+                type = types.lazyAttrsOf types.package;
                 readOnly = true;
                 description = ''
-                  The `hlsCheck` derivation generated for this project.
+                  The flake checks generated for this project.
                 '';
               };
+
             };
           };
           projectSubmodule = types.submoduleWith {
@@ -310,9 +317,8 @@ in
               checks =
                 mergeMapAttrs
                   (name: project:
-                    lib.optionalAttrs (project.autoWire && project.devShell.enable && project.devShell.hlsCheck.enable) {
-                      "${name}-hls" = project.outputs.hlsCheck;
-                    })
+                    project.outputs.checks
+                  )
                   config.haskellProjects;
             };
         });